requirements.txt
Django
djangorestframework>=3.14,<4.0
celery>=5.3,<6.0
redis>=4.5,<5.0
beautifulsoup4>=4.11,<5.0
selenium>=4.15,<5.0
requests>=2.31,<3.0
httpx>=0.25,<1.0
scrapy>=2.11,<3.0
aiohttp>=3.9,<4.0
builtwith>=1.3,<2.0
playwright>=1.40,<2.0
fake-useragent>=1.4,<2.0
ratelimit>=2.2,<3.0
tqdm>=4.65,<5.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
lxml>=4.9,<5.0
html5lib>=1.1,<2.0
cssutils>=2.7,<3.0
jsbeautifier>=1.14,<2.0
whitenoise>=6.6,<7.0
gunicorn>=21.2,<22.0
psycopg2-binary>=2.9,<3.0  # For PostgreSQL, optional
django-redis>=5.3,<6.0
django-filter>=23.5,<25.0
drf-spectacular>=0.27,<1.0
pandas>=2.1,<3.0
diff-match-patch>=20230430,<20240000
channels>=4.0,<5.0  # For WebSockets
channels-redis>=4.1,<5.0
